refactor: Move http clients related to register to new namespace
