+++ src/Altinn.App.Core/Infrastructure/Register/HttpClients/RegisterERClient.cs
@@ -1,7 +1,4 @@
-using System;
-using System.Net.Http;
 using System.Net.Http.Headers;
-using System.Threading.Tasks;
 using Altinn.App.PlatformServices.Extensions;
 using Altinn.App.Services.Configuration;
 using Altinn.App.Services.Constants;
@@ -13,7 +10,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 
-namespace Altinn.App.Services.Implementation
+namespace Altinn.App.Core.Infrastructure.Register.HttpClients
 {
     /// <summary>
     /// A client for retrieving ER data from Altinn Platform.
